跟单套利bot/requirements.txt
py-clob-client>=0.40.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
